backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
pypdf>=4.0.0
pytesseract>=0.3.0
pillow>=10.0.0
pdf2image>=1.17.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
slowapi>=0.1.0
spacy>=3.7.0